Refactor befor code review
