+++ frontend/src/components/shared/layouts/footer.tsx
