+++ frontend/src/features/users/components/updateUserForm.tsx
@@ -7,9 +7,9 @@ import type { ChangeEventHandler, FormEventHandler, FC } from "react";
 import { createUserUpdateAction } from "@/features/users/stores/reducers/updateUserReducer";
 import { useAppDispatch, useAppSelector } from "@/stores/hooks";
 import type { SignupInput as UpdateUserInput } from "@/features/users/types/authTypes";
-import { InputField } from "@/components/shared/inputField";
-import { SubmitButton } from "@/components/shared/submitButton";
-import { BlueButton, GreenButton } from "@/components/shared/buttons";
+import { InputField } from "@/components/shared/form-elements/inputField";
+import { SubmitButton } from "@/components/shared/buttons/submitButton";
+import { BlueButton, GreenButton } from "@/components/shared/buttons/buttons";
 
 const UpdateUserForm: FC = () => {
   const dispatch = useAppDispatch();
